chore(Benchmark Scripts): Extracted common methods into utils class
